backend/requirements.txt
Flask==2.2.5
flask-cors==4.0.0
Flask-SQLAlchemy==3.0.5
pandas==1.4.3